+++ src/app/services/auth.service.ts
@@ -63,15 +63,7 @@ export class AuthService {
     //return this.http.get(`${this.apiConstants.apiUrl}${this.apiConstants.resources}/resources`, _options);
     return this.http.get(`${this.apiConstants.apiUrl}${this.apiConstants.resources}/stateInfo`, _options);
   }
-  getPoliciesList() {
-    const _options = {
-      headers: new HttpHeaders({
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${localStorage.getItem('token')}`
-      })
-    };
-    return this.http.get(`${this.apiConstants.apiUrl}${this.apiConstants.policiesList}`, _options);
-  }
+
   getResourcesFast() {
     const _options = {
       headers: new HttpHeaders({
